Converted the resource component interface into a class that professional and instructional inherit.  Moved shared functionality to the resource component.  Moved shared components out of the instructional folder.
