+++ webapp/src/app/resource/components/attachments/attachments.component.scss
