+++ webapp/src/app/resource/components/formative/formative.component.scss
