+++ webapp/src/app/resource/components/step-by-step/step-by-step.component.scss
