+++ webapp/src/app/resource/instructional/instructional-resource.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { mockDataServiceProviders } from 'src/app/app.module.spec';
 import { mockResourceModel } from 'src/app/data/mock-data';
-import { ResourceComponent } from '../resource-component.interface';
+import { ResourceComponent } from '../resource.component';
 import { ResourceModule } from '../resource.module';
 import { InstructionalResourceComponent } from './instructional-resource.component';
