+++ webapp/src/app/resource/professional/professional-resource.component.ts
@@ -1,18 +1,14 @@
-import { Component, OnInit } from '@angular/core';
-import { ResourceComponent } from '../resource-component.interface';
-import { ResourceModel } from 'src/app/data/resource/model/resource.model';
+import { ChangeDetectorRef, Component } from '@angular/core';
+import { DomSanitizer } from '@angular/platform-browser';
+import { ResourceComponent } from '../resource.component';
 
 @Component({
   selector: 'sbdl-professional-resource',
   templateUrl: './professional-resource.component.html',
-  styleUrls: ['./professional-resource.component.scss']
+  styleUrls: ['./../resource.component.scss']
 })
-export class ProfessionalResourceComponent implements OnInit, ResourceComponent {
-  model: ResourceModel;
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
+export class ProfessionalResourceComponent extends ResourceComponent {
+  constructor(cdRef: ChangeDetectorRef, sanitizer: DomSanitizer) {
+    super(cdRef, sanitizer);
+   }
 }
