--- webapp/src/app/resource/resource-component.interface.ts
@@ -1,5 +1,0 @@
-import { ResourceModel } from '../data/resource/model/resource.model';
-
-export interface ResourceComponent {
-    model: ResourceModel;
-}
