+++ webapp/src/app/resource/resource.component.scss
